requirements.txt
Faker==14.2.0
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
python-dotenv==0.21.0
webargs==8.2.0
